+++ processor/src/host/mod.rs
@@ -2,8 +2,8 @@ use alloc::{sync::Arc, vec::Vec};
 use core::future::Future;
 
 use miden_core::{
-    Felt, Word,
-    advice::AdviceMap,
+    Word,
+    advice::{AdviceMap, AdviceStack},
     crypto::merkle::InnerNodeInfo,
     events::{EventId, EventName},
 };
@@ -29,14 +29,14 @@ pub use mast_forest_store::{LoadedMastForest, MastForestStore, MemMastForestStor
 /// Any possible way an event can modify the advice provider.
 #[derive(Debug, PartialEq, Eq)]
 pub enum AdviceMutation {
-    ExtendStack { values: Vec<Felt> },
+    ExtendStack { stack: AdviceStack },
     ExtendMap { other: AdviceMap },
     ExtendMerkleStore { infos: Vec<InnerNodeInfo> },
 }
 
 impl AdviceMutation {
-    pub fn extend_stack(iter: impl IntoIterator<Item = Felt>) -> Self {
-        Self::ExtendStack { values: Vec::from_iter(iter) }
+    pub fn extend_advice_stack(stack: AdviceStack) -> Self {
+        Self::ExtendStack { stack }
     }
 
     pub fn extend_map(other: AdviceMap) -> Self {